requirements.txt
Click>=7.0
flask>=1.0.2
blinker>=1.4
flask-jwt-extended>=3.14.0
cryptography>=2.3.1
itsdangerous>=0.24
flask-caching>=1.4.0
cherrypy>=18.0.1
wsgi-request-logger>=0.4.6
Flask-SQLAlchemy>=2.3.2
Flask-Login>=0.4.1
Flask-JWT-Extended>=3.14.0
flask-marshmallow>=0.9.0
simplejson>=3.15.0
sqlalchemy>=1.2.15
sqlalchemy-utils>=0.33.9
sqlalchemy-json>=0.2.1
grpcio>=1.17.1
chaosplatform-grpc>=0.1.0
chaosplatform-relational-storage>=0.2.0
toml>=0.10.0
croniter>=0.3.27
